requirements.txt
Django==3.2.18
dj-database-url==1.2.0
djangorestframework==3.14.0
environs==9.5.0
psycopg2-binary==2.9.5
rosreestr2coord==4.2.7